Move files into new-simulation directory
