+++ src/components/new-simulation/simulation-results.scss
@@ -1,4 +1,4 @@
-@use "./vars.scss";
+@use "../vars.scss";
 
 .graphs-header {
   margin-bottom: 26px;
